Extracted class to separate namespace
